DeskClock: Always use dark theme colors
There's no real distinction in the clock app to begin with — the "light"
theme is still dark. The difference is that the dark theme uses a fairly
neutral color palette with a desaturated accent color as per Material
guidelines, while the light theme has saturated colors for the entire UI
(including background, of all things) that do not fit well here.

Change-Id: I9a9ccfd911d069f556ae2fd335539def20e022bc